feat(asset-proxy): handle errors with placeholder and remove sentry logging
