+++ src/embed.ts
@@ -1,6 +1,13 @@
 import * as t from 'io-ts'
 
-import { SentryFactory, SentryReporter, responseToContext, Url } from './utils'
+import {
+  SentryFactory,
+  SentryReporter,
+  responseToContext,
+  Url,
+  getPlaceholder,
+  isImageResponse,
+} from './utils'
 
 export async function embed(
   request: Request,
@@ -270,25 +277,4 @@ async function getWikimediaThumbnail(url: URL) {
   if (isImageResponse(imgRes)) return imgRes
 
   return getPlaceholder()
-}
-
-function getPlaceholder() {
-  const placeholderBase64 =
-    'iVBORw0KGgoAAAANSUhEUgAAAwAAAAGwAQMAAAAkGpCRAAAAA1BMVEXv9/t0VvapAAAAP0lEQVR42u3BMQEAAADCIPuntsUuYAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAQOqOwAAHrgHqAAAAAAElFTkSuQmCC'
-  const placeholder = Uint8Array.from(atob(placeholderBase64), (c) =>
-    c.charCodeAt(0),
-  )
-  return new Response(placeholder, {
-    status: 200,
-    statusText: 'OK',
-    headers: {
-      'Content-Type': 'image/png',
-      'Content-Length': placeholder.length.toString(),
-    },
-  })
-}
-
-function isImageResponse(res: Response): boolean {
-  const contentType = res.headers.get('content-type') ?? ''
-  return res.status === 200 && contentType.startsWith('image/')
 }
