+++ src/utils/index.tsx
@@ -4,3 +4,4 @@ export * from './cf-environment'
 export * from './url'
 export * from './ui'
 export * from './cache'
+export * from './image'
